% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/131_computegroup.R
\name{fn_computegroup}
\alias{fn_computegroup}
\title{fn_computegroup}
\usage{
fn_computegroup(data, scales, siiparams)
}
\arguments{
\item{data}{the data send by the geom_class, the result of the integration of user supplied data and expanded structure .}

\item{scales}{dummy text}

\item{siiparams}{The (adjusted) geom parameterset, forwarded to lower level functions as siiparams}
}
\value{
a dataframe which contains transformed or enriched data, usable for plotting
}
\description{
fn_computegroup
}
