% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/136_computegroup_subA.R
\name{fn_cornerpoints}
\alias{fn_cornerpoints}
\title{fn_cornerpoints}
\usage{
fn_cornerpoints(df, siiparams)
}
\arguments{
\item{df}{The dataframe resulted from last step}

\item{siiparams}{dummy text}
}
\value{
a dataframe with 5 added columns (ri, ro, db, de, plottype)
}
\description{
fn_cornerpoints
}
