% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/121_setupdata_risksurface.R
\name{fn_determinelevels}
\alias{fn_determinelevels}
\title{fn_determinelevels reduces a vector with levels to a list with unique items, possible after further selection or transformation to value}
\usage{
fn_determinelevels(vector_in, ind_value = FALSE, ind_integer = FALSE,
  ind_d = TRUE, ind_o = TRUE, ind_onlyspecials = FALSE)
}
\arguments{
\item{vector_in}{a vector with levels, this function reduces it to unique values}

\item{ind_value}{default = FALSE: levels are returned as value. Implies that ind_d and ind_o are set to FALSE (while default is TRUE)}

\item{ind_integer}{default = FALSE: levels are returned as value. Implies that ind_value is set to TRUE, and hence ind_d and ind_o to FALSE}

\item{ind_d}{default = TRUE: xxxxd levels are included in the result}

\item{ind_o}{default = TRUE: xxxxo levels are included in the result}

\item{ind_onlyspecials}{default = FALSE: non xxxxd/xxxo levels are NOT included in the result}
}
\value{
a vector of levels, in character, numeric or integer format
}
\description{
fn_determinelevels reduces a vector with levels to a list with unique items, possible after further selection or transformation to value
}
