% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/131_computegroup.R
\name{fn_geomsiidatatopoints}
\alias{fn_geomsiidatatopoints}
\title{fn_geomsiidatatopoints}
\usage{
fn_geomsiidatatopoints(df, siiparams)
}
\arguments{
\item{df}{dummy text}

\item{siiparams}{dummy text}
}
\value{
a list with one item 'df' (data.frame)
}
\description{
fn_geomsiidatatopoints
}
