% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/111_setupparams.R
\name{fn_levelonedescription}
\alias{fn_levelonedescription}
\title{fn_levelonedescription extracts the name of level 1 item from the structure}
\usage{
fn_levelonedescription(params)
}
\arguments{
\item{params}{the params send by the geom_class, possible adjusted by setup_params.}
}
\value{
the value maxscrvalue, whether or not this is provided by the user
}
\description{
fn_levelonedescription extracts the name of level 1 item from the structure
}
