% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/111_setupparams.R
\name{fn_maxscrvalue}
\alias{fn_maxscrvalue}
\title{fn_maxscrvalue
all elements in the resulting dataset are scaled to surface of value of the largest level 1 (SCR) element
 in the set, unless the value to be scaled to is provided as a parameter.}
\usage{
fn_maxscrvalue(data, params)
}
\arguments{
\item{data}{the data send by the geom_class, the result of the integration of user supplied data and expanded structure .}

\item{params}{the params send by the geom_class, possible adjusted by setup_params.}
}
\value{
the value maxscrvalue, whether or not this is provided by the user.
}
\description{
fn_maxscrvalue
all elements in the resulting dataset are scaled to surface of value of the largest level 1 (SCR) element
 in the set, unless the value to be scaled to is provided as a parameter.
}
