% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/136_computegroup_subA.R
\name{fn_polygonpoints}
\alias{fn_polygonpoints}
\title{fn_polygonpoints}
\usage{
fn_polygonpoints(df, siiparams, counter_polyorder)
}
\arguments{
\item{df}{The dataframe resulted from last step}

\item{siiparams}{The (adjusted) geom parameterset, forwarded to lower level functions as siiparams}

\item{counter_polyorder}{a continues counter, needed for sorting polygonpoints}
}
\value{
a list with two items (df (dataframe), counter_polyorder (numeric, updated)
}
\description{
fn_polygonpoints
}
