% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/123_setupdata_riskconnection.R
\name{fn_setupdata_connection}
\alias{fn_setupdata_connection}
\title{fn_setupdata_connection}
\usage{
fn_setupdata_connection(data)
}
\arguments{
\item{data}{the data send by the geom_class, due to the simple nature of the desired plot no structure information is needed.}
}
\value{
an adjusted version of input dataframe data
}
\description{
fn_setupdata_connection
}
