% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/121_setupdata_risksurface.R
\name{fn_setupdata_surfaces}
\alias{fn_setupdata_surfaces}
\title{fn_setupdata_surfaces}
\usage{
fn_setupdata_surfaces(data, params)
}
\arguments{
\item{data}{the data send by the geom_class, the result of the integration of user supplied data and expanded structure .}

\item{params}{the params send by the geom_class, possible adjusted by setup_params.}
}
\value{
an adjusted version of input dataframe data
}
\description{
fn_setupdata_surfaces
}
