% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/137_computegroup_subB.R
\name{fn_sqdegseq}
\alias{fn_sqdegseq}
\title{fn_sqdegseq gives a sequence of degrees based on a starting and end rotation, with al the multiples of 45 degrees within the range. If starting degrees > enddegrees then (multples of) 360 degrees are added until start degrees is smaller then enddegrees.}
\usage{
fn_sqdegseq(startdegrees, enddegrees, counterclockwise = FALSE,
  reset = FALSE)
}
\arguments{
\item{startdegrees}{a startingpoint of degrees (as used in a compass)}

\item{enddegrees}{the endpoint of degrees (as used in a compass)}

\item{counterclockwise}{(bln, default = FALSE)}

\item{reset}{(bln, default = FALSE) if reset is True then startdegrees is reset to a number between 0 and 360, and enddegrees to a number between 0 and 720 degrees}
}
\value{
a sequence of at least two compass directions, with the multiples of 45 degrees in between the outer limits.
}
\description{
fn_sqdegseq gives a sequence of degrees based on a starting and end rotation, with al the multiples of 45 degrees within the range. If starting degrees > enddegrees then (multples of) 360 degrees are added until start degrees is smaller then enddegrees.
}
