% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/136_computegroup_subA.R
\name{fn_squareconversion}
\alias{fn_squareconversion}
\title{fn_squareconversion}
\usage{
fn_squareconversion(df, siiparams)
}
\arguments{
\item{df}{a dataframe with columns ri, ro, db, de}
}
\value{
the same dataframe with ri, ro, db and de converted in parameters for a square plot
}
\description{
fn_squareconversion
}
