% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/136_computegroup_subA.R
\name{fn_squareddegrees}
\alias{fn_squareddegrees}
\title{fn_squareddegrees The actual transformation of a part of a circle to the (surfacewise)  equivalent of the part of a square}
\usage{
fn_squareddegrees(circledegrees)
}
\arguments{
\item{circledegrees}{(no default) an amount in degrees}
}
\value{
an amount in compass degrees
}
\description{
fn_squareddegrees The actual transformation of a part of a circle to the (surfacewise)  equivalent of the part of a square
}
