% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/137_computegroup_subB.R
\name{fn_sqXP}
\alias{fn_sqXP}
\title{fn_sqXP calculating the horizontal offset of the point, based on a reference value (vertdistance) and compass direction.}
\usage{
fn_sqXP(vertdistance, degrees, xsign, rounding)
}
\arguments{
\item{vertdistance}{dummy text}

\item{degrees}{dummy text}

\item{xsign}{dummy text}

\item{rounding}{dummy text}
}
\value{
a distance
}
\description{
fn_sqXP calculating the horizontal offset of the point, based on a reference value (vertdistance) and compass direction.
}
