% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/137_computegroup_subB.R
\name{fn_sqYP}
\alias{fn_sqYP}
\title{fn_sqYP calculating the vertical offset of the point, based on a reference value (vertdistance) and compass direction.}
\usage{
fn_sqYP(vertdistance, degrees, ysign, rounding)
}
\arguments{
\item{vertdistance}{dummy text}

\item{degrees}{dummy text}

\item{ysign}{dummy text}

\item{rounding}{dummy text}
}
\value{
a distance
}
\description{
fn_sqYP calculating the vertical offset of the point, based on a reference value (vertdistance) and compass direction.
}
