% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/121_setupdata.R
\name{fn_structure_data_integration}
\alias{fn_structure_data_integration}
\title{fn_structure_data_integration}
\usage{
fn_structure_data_integration(expandedstructure, data)
}
\arguments{
\item{expandedstructure}{result of fn_structure_expansion()}

\item{data}{:dataframe in tidyverse format with (minimal) columns id, x, y, description, value and possible columns comparewithid, COLOR, FILLCOLOR, PANEL}
}
\value{
data
}
\description{
fn_structure_data_integration
}
