% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/121_setupdata.R
\name{fn_structure_expansion}
\alias{fn_structure_expansion}
\title{fn_structure_expansion}
\usage{
fn_structure_expansion(params)
}
\arguments{
\item{params}{dummy text}
}
\value{
a dataframe
}
\description{
fn_structure_expansion
}
