% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/131_computegroup.R
\name{fn_transform_outlinedf}
\alias{fn_transform_outlinedf}
\title{fn_transform_outlinedf}
\usage{
fn_transform_outlinedf(df)
}
\arguments{
\item{df}{a dataframe with column level and some other levels}
}
\value{
a tidyverse dataframe
}
\description{
fn_transform_outlinedf
}
