% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/131_computegroup.R
\name{fn_transform_plotdetails}
\alias{fn_transform_plotdetails}
\title{fn_transform_plotdetails}
\usage{
fn_transform_plotdetails(df, siiparams)
}
\arguments{
\item{df}{a dataframe with column level and some other levels}

\item{siiparams}{the (extended) parameterset}
}
\value{
a tidyverse dataframe
}
\description{
fn_transform_plotdetails
}
