% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/137_computegroup_subB.R
\name{fnCirclepoints}
\alias{fnCirclepoints}
\title{fnCirclepoints}
\usage{
fnCirclepoints(ri, ro, db, de, plottype, indication11, indication13,
  stepsize = 0.5)
}
\arguments{
\item{ri}{the  inner radius of the circle segment}

\item{ro}{the outer radius of the circle segment}

\item{db}{degrees beginning of section (compaswise: 0 degrees is north, 90 degrees is east)}

\item{de}{degrees end of section}

\item{plottype}{"circle" of "outer", the former only requiring a ro}

\item{indication11}{dummy text}

\item{indication13}{dummy text}

\item{stepsize}{default value = 0.5, as of now no calls to this function override this default. the stepsize in degrees along the inner or outer lines,}
}
\value{
a dataframe with points to be plotted, relative from (0,0) coordinates system.
The dataframe has 3 columns: xpoint, ypoint, outlinetype (from puntenfunctie omschrijving/description)
## outlinetype are 1 through 4 and 11 and 13.
      type 1 : line from point defined by ri/db  to ro/db (radius outward)
      type 2 : line from point defined by ro/db  to ro/de (outer line of section)
      type 3 : line from point defined by ro/de  to ri/de (radius inward)
      type 4 : line from point defined by ri/de  to ri/db  outer line of section)
}
\description{
fnCirclepoints
}
