% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/131_computegroup.R
\name{fnComputegroup}
\alias{fnComputegroup}
\title{fnComputegroup}
\usage{
fnComputegroup(data, scales, siiparams)
}
\arguments{
\item{data}{dummy text}

\item{scales}{dummy text}

\item{siiparams}{dummy text}
}
\value{
a dataframe which containes transformed or enriched data, usable for plotting
}
\description{
fnComputegroup
}
