% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/136_computegroup_subA.R
\name{fnCornerpoints}
\alias{fnCornerpoints}
\title{fnCornerpoints}
\usage{
fnCornerpoints(df, siiparams)
}
\arguments{
\item{df}{a dataframe}

\item{siiparams}{dummy text}
}
\value{
a dataframe with 5 added columns (ri,ro,db,de,plottype)
}
\description{
fnCornerpoints
}
