% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/136_computegroup_subA.R
\name{fnRotation}
\alias{fnRotation}
\title{fnRotation}
\usage{
fnRotation(df, siiparams)
}
\arguments{
\item{df}{The dataframe resulted from last step}

\item{siiparams}{The (adjusted) geom parameterset, forwarded to lower level functions as siiparams}
}
\value{
a dataframe with adjusted db and de columns
}
\description{
fnRotation
}
