% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/122_setupdataoutline.R
\name{fnSetupdata_outline}
\alias{fnSetupdata_outline}
\title{fnSetupdata_outline}
\usage{
fnSetupdata_outline(data, params)
}
\arguments{
\item{data}{the data send by the geom_class}

\item{params}{the params send by the geom_classs}
}
\value{
an adjusted version of input dataframe data
}
\description{
fnSetupdata_outline
}
