% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/121_setupdata.R
\name{fnSetupdata_surfaces}
\alias{fnSetupdata_surfaces}
\title{fnSetupdata_surfaces}
\usage{
fnSetupdata_surfaces(data, params)
}
\arguments{
\item{data}{the data send by the geom_class}

\item{params}{the params send by the geom_classs}
}
\value{
an adjusted version of input dataframe data
}
\description{
fnSetupdata_surfaces
}
