% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/136_computegroup_subA.R
\name{fnSquareconversion}
\alias{fnSquareconversion}
\title{fnSquareconversion}
\usage{
fnSquareconversion(df)
}
\arguments{
\item{df}{a dataframe with colums ri, ro, db, de}
}
\value{
the same dataframe with ri, ro, db and de converted in parameters for a square plot
}
\description{
fnSquareconversion
}
