% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/137_computegroup_subB.R
\name{fnsqYP}
\alias{fnsqYP}
\title{fnsqYP calculating the vertical offset of the point, based on a reference value (vertdistance) and compass direction.}
\usage{
fnsqYP(vertdistance, degrees, ysign, rounding)
}
\arguments{
\item{vertdistance}{dummy text}

\item{degrees}{dummy text}

\item{ysign}{dummy text}

\item{rounding}{dummy text}
}
\value{
a distance
}
\description{
fnsqYP calculating the vertical offset of the point, based on a reference value (vertdistance) and compass direction.
}
