% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101_geomandstat.R
\name{geom_siiconnection}
\alias{geom_siiconnection}
\title{geom_siiconnection}
\usage{
geom_siiconnection(data = NULL, mapping = NULL, stat = "siiconnection",
  position = "identity", na.rm = FALSE, show.legend = NA,
  inherit.aes = TRUE, ...)
}
\arguments{
\item{data}{the dataset in tidyverse format (column 'description' as a factor). see examples in \code{\link{sii_z_ex2_data}} or \code{\link{sii_z_ex1_data}}}

\item{mapping}{required aes(thetics) : x (i.e. time, longitude, integer), y (i.e SCR ratio, lattitude), id, description (), value and comparewithid}

\item{stat}{default stat is statSiiconnection, combinations with other stat's are not tested}

\item{position}{standard ggplot function}

\item{na.rm}{standard ggplot function}

\item{show.legend}{standard ggplot function}

\item{inherit.aes}{standard ggplot function}

\item{...}{ellipsis, a standard R parameter}
}
\value{
a ggplot object
}
\description{
Plots a line between those datapoints which have a matching value in the columns 'id' and 'comparewithid'.
}
\examples{
\dontrun{
dummy
}
}
