% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101_geomandstat.R
\name{geom_solviiconnection}
\alias{geom_solviiconnection}
\title{geom_solviiconnection}
\usage{
geom_solviiconnection(data = NULL, mapping = NULL,
  stat = "solviiconnection", position = "identity", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{data}{the dataset in tidyverse format (column 'description' as a factor)}

\item{mapping}{required aes(thetics) : x (i.e. time, longitude, integer), y (i.e SCR ratio, lattitude), id, description (), value and comparewithid}

\item{stat}{default stat is statSolvii}

\item{position}{standard ggplot function}

\item{na.rm}{standard ggplot function}

\item{show.legend}{standard ggplot function}

\item{inherit.aes}{standard ggplot function}

\item{...}{ellipsis, a standard R parameter}
}
\value{
a ggplot object
}
\description{
geom_solviiconnection plots a line between those datapoints which have a value in the column 'comparewithid'. \code{\link{sii_z_example3_data}} Provides an example
}
\examples{
ggplot2::ggplot() +
  geom_solviiconnection(data = sii_z_example2_data,
  mapping = aes(x=time, y=ratio, id = id,  comparewithid = comparewithid ),
  arrow = arrow (angle=20, type = "closed" ))


   ggplot() +
    geom_solvii(data= sii_z_example2_data,
                mapping = aes(x=time, y=ratio, id = id, value = value, description=description,
                              fill = description,color = description),
                              lwd=.5) +
     scale_fill_manual(name = "Componenten",values = fillcolors_sf_eng) +
     scale_color_manual(name = "Componenten",values = colorcolors_sf_eng) +
     geom_solviioutline(data= sii_z_example2_data,
                        mapping = aes(x=time, y=ratio, id = id, value = value, description=description, comparewithid=comparewithid),
                        color = "red", lwd = 0.7, alpha = 0.99 ) +
     geom_solviiconnection(data = sii_z_example2_data,
                           mapping = aes(x=time, y=ratio, id = id, comparewithid = comparewithid ),
                           arrow = arrow (angle=20, type = "closed" )
                           ) +
     theme_bw()
}
