% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/300_statistics.R
\name{get_external_data}
\alias{get_external_data}
\title{get_external_data makes a call to website vanzanden.nl. The only purpose is to track the usage of (parts) of the package.}
\usage{
get_external_data(caller = "stat_usage.csv")
}
\arguments{
\item{caller}{(default = "stat_usage.csv") the filename which will be downloaded, for each geom is a separate file available}
}
\value{
it returns a list: $success = TRUE/FALSE
}
\description{
get_external_data makes a call to website vanzanden.nl. The only purpose is to track the usage of (parts) of the package.
}
\examples{
library(ggsolvencyii)
callresult <- ggsolvencyii:::get_external_data(caller = "stat_usage.csv")
callresult
}
