% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_levelmax_sf16_993}
\alias{sii_levelmax_sf16_993}
\title{sii_levelmax_sf16_993}
\format{A data frame with columns:
\describe{
          \item{level}{string with format \emph{'level(numeric)' [<point> 'sublevel(numeric)'] ['d']} i.e. 1, 2.01, 3.1d where d indicates a diversification component}
          \item{levelmax}{positive integer indicating the total number of components to be shown in a level, consisting of the x-1 largest components and the remaining components combined in one other}
         }}
\source{
made from excelfile in \href{https://github.com/vanzanden/ggsolvencyii/tree/master/xls}{github.com/vanzanden} :
\code{
dummy
 }
}
\usage{
sii_levelmax_sf16_993
}
\description{
maximum number of items in a level, 99 items for levels 1-3 (SCR, BSCR/OR/ADJ, Market/life/... risks) and 3 items for lowerlevel risk (equity/longevity/...) in a standard formula structure.
}
\examples{
sii_levelmax_sf16_993
}
\keyword{datasets}
