% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_levelmax_sf16_995}
\alias{sii_levelmax_sf16_995}
\title{sii_levelmax_sf16_995}
\format{A data frame with columns:
\describe{
          \item{level}{string with format 'level(numeric)' [<point> 'sublevel'] ['d'] } with d indicating diversification
          \item{levelmax}{integer \ code indicating the total number of components to be shown in a level, consisting of the x-1 largest components and the remaining components combined in one other}
         }}
\usage{
sii_levelmax_sf16_995
}
\description{
tekst
}
\examples{
sii_levelmax_sf16_995
}
\keyword{datasets}
