% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_outline2_sf16_nld}
\alias{sii_outline2_sf16_nld}
\title{sii_outline2_sf16_nld}
\format{A data frame with columns:
\describe{
           \item{levelordescription}{string with format \emph{'level(numeric)' [<point> 'sublevel(numeric)'] ['d']} i.e. 1, 2.01, 3.1d where d indicates a diversification component \strong{AND/OR} description(s) from the corresponding structure }
           \item{outline1}{boolean (logical): the radial line }
           \item{outline2}{boolean (logical): the outer line  }
           \item{outline3}{boolean (logical): the radial line  }
           \item{outline4}{boolean (logical): the inner line }
           \item{outline11}{NOT YET IMPLEMENTED: boolean (logical): analogue to outline1, but but only for the first plotted component of that level. The components in each  level are plotted clockwise}
           \item{outline13}{NOT YET IMPLEMENTED: boolean (logical): analogue to outline3, but but only for the last plotted component of that level. }
         }}
\source{
determined based on specification of EIOPA standard formula Solvency II structure
  made from excelfile in \href{https://github.com/vanzanden/ggsolvencyii/tree/master/xls}{github.com/vanzanden}, from there transfered to R environment with code in \code{preparingtables.R}.
}
\usage{
sii_outline2_sf16_nld
}
\description{
A table for \code{\link{geom_sii_riskoutline}} indicating which outlines of each item should be shown, specified per level and/or description. the latter overrule the former. \cr
when defining an item (or the \code{squared = TRUE} transformation) 4 lines can be distinguised, a radialline going outwards, a circle segment (clockwise), a radialline going inwards, a circle segment (counterclockwise). These are numbered as outline1 to outline4.
}
\examples{
sii_outline2_sf16_nld
}
\keyword{datasets}
