% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/300_statistics.R
\name{sii_setoption_statistics_httpfailurecounter}
\alias{sii_setoption_statistics_httpfailurecounter}
\title{sii_setoption_statistics_httpfailurecounter}
\usage{
sii_setoption_statistics_httpfailurecounter(cnt = 0)
}
\arguments{
\item{cnt}{(default = 0) the value to be assigned to the counter}
}
\value{
nothing
}
\description{
sii_setoption_statistics_httpfailurecounter
}
\examples{
# library(ggsolvencyii)
ggsolvencyii:::sii_setoption_statistics_httpfailurecounter()
ggsolvencyii:::sii_setoption_statistics_httpfailurecounter(cnt = 10)

}
