% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/300_statistics.R
\name{sii_setoption_statistics_remind}
\alias{sii_setoption_statistics_remind}
\title{sii_setoption_statistics_remind
sets the option to remind that participation is set to TRUE (if that is the case) after 1,2,3,5,10,50,100,200,500,1000,1500,2500 calls}
\usage{
sii_setoption_statistics_remind(remind = TRUE, quiet = FALSE)
}
\arguments{
\item{remind}{(default = TRUE) : whether to remind that participation is set}

\item{quiet}{(default = FALSE) : whether to confirm that the option is set}
}
\value{
remind
}
\description{
sii_setoption_statistics_remind
sets the option to remind that participation is set to TRUE (if that is the case) after 1,2,3,5,10,50,100,200,500,1000,1500,2500 calls
}
\examples{
library(ggsolvencyii)
sii_setoption_statistics_remind()
sii_setoption_statistics_remind(remind = FALSE)
sii_setoption_statistics_remind(remind = FALSE, quiet = TRUE)

}
