% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/300_statistics.R
\name{sii_setoption_statistics_remindercounter}
\alias{sii_setoption_statistics_remindercounter}
\title{sii_setoption_statistics_remindercounter}
\usage{
sii_setoption_statistics_remindercounter(cnt = 0)
}
\arguments{
\item{cnt}{(default = 0)}
}
\value{
nothing
}
\description{
sii_setoption_statistics_remindercounter
}
\examples{
# library(ggsolvencyii)
ggsolvencyii:::sii_setoption_statistics_remindercounter()
ggsolvencyii:::sii_setoption_statistics_remindercounter(cnt = 10)

}
