% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_x_edgecolors_sf16_nld}
\alias{sii_x_edgecolors_sf16_nld}
\title{sii_x_edgecolors_sf16_nld}
\format{A list with items in the form "\emph{description}"= "#\emph{colorcode}":}
\source{
made with \code{preparingtables.R} (not included in this package) which can be found on \href{https://github.com/vanzanden/ggsolvencyii/tree/master/_data_helperfiles}{github.com/vanzanden/ggsolvencyii}
}
\usage{
sii_x_edgecolors_sf16_nld
}
\description{
a colorset, as named list, where each risk has the edgecolor of the "motherlevel", as defined in the \code{..fillcolor..} variable.
}
\examples{
sii_x_edgecolors_sf16_nld
}
\keyword{datasets}
