% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_x_fillcolors_sf16_eng}
\alias{sii_x_fillcolors_sf16_eng}
\title{sii_x_fillcolors_sf16_eng}
\format{A list with items in the form "\emph{description}"= "#\emph{colorcode}":}
\source{
made with \code{preparingtables.R} (not included in this package) which can be found on \href{https://github.com/vanzanden/ggsolvencyii/tree/master/xls}{github.com/vanzanden/ggsolvencyii}
}
\usage{
sii_x_fillcolors_sf16_eng
}
\description{
a colorset where each accumulationlevel and each type of risk has it's own color, where ".._other" has the same color as the "motherlevel", i.e. 'market_other' has the same color as 'market' while other 'm_..' risks have a different hue.
}
\examples{
sii_x_fillcolors_sf16_eng
}
\keyword{datasets}
