% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_z_ex1_structure}
\alias{sii_z_ex1_structure}
\title{sii_z_ex1_structure}
\format{A data frame with columns:
\describe{
          \item{description}{string: the name of the component, or a diversification}
          \item{level}{string with format \emph{'level(numeric)' [<point> 'sublevel(numeric)'] ['d']} i.e. 1, 2.01, 3.1d where d indicates a diversification component}
          \item{childlevel}{string of the same format as level, indicating the level of which the components make up this item }
         }}
\source{
determined based on specification of EIOPA standard formula Solvency II structure
  made from excelfile in \href{https://github.com/vanzanden/ggsolvencyii/tree/master/_data_helperfiles}{github.com/vanzanden}, from there transfered to R environment with code in \code{preparingtables.R}.
}
\usage{
sii_z_ex1_structure
}
\description{
a adapted (compact) structure, for a life-insurer with only  market and life risk and their subrisks
}
\examples{
sii_z_ex1_structure
}
\keyword{datasets}
