% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_z_ex2_data}
\alias{sii_z_ex2_data}
\title{sii_z_ex2_data}
\format{A data frame with columns:
\describe{
           \item{time}{numerical: a representation of an x value: \cr The x aesthetic could also be a normal numbering, or a longitude}
           \item{ratio}{numerical: solvency II ratio, a representation of an y value: \cr y aesthetic could also be a lattitude}
           \item{description}{character: component of the structure}
           \item{id}{ number: }
           \item{value}{numerical: }
           \item{comparewithid}{numerical: }
         }}
\source{
loosely based on public SFCR report of a medium sized dutch life insurer
  made from excelfile in \href{https://github.com/vanzanden/ggsolvencyii/tree/master/xls}{github.com/vanzanden}, from there transfered to R environment with code in \code{preparingtables.R}.
}
\usage{
sii_z_ex2_data
}
\description{
example dataset, in tidyverse format, with all columns of the complete SF16 structure present in the data but with only a few risks filled with non-zero values,
}
\examples{
sii_z_ex2_data
}
\keyword{datasets}
