% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_z_ex6_levelmax}
\alias{sii_z_ex6_levelmax}
\title{sii_z_ex6_levelmax}
\format{A data frame with columns:}
\source{
made from excelfile in \href{https://github.com/vanzanden/ggsolvencyii/tree/master/data_helperfiles}{github.com/vanzanden}, from there transfered to R environment with code in \code{preparingtables.R}.
}
\usage{
sii_z_ex6_levelmax
}
\description{
tekst
}
\examples{
sii_z_ex6_levelmax
}
\keyword{datasets}
