% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_z_ex8_map}
\alias{sii_z_ex8_map}
\title{sii_z_ex8
#'
A map of western Europe, to show a location based geom_sii_risk...() plot}
\format{a ggplot2 object}
\source{
stamen
  made with code in \code{preparingmap.R}, located in \href{https://github.com/vanzanden/ggsolvencyii/tree/master/_data_helperfiles}{github.com/vanzanden}.
}
\usage{
sii_z_ex8_map
}
\description{
sii_z_ex8
#'
A map of western Europe, to show a location based geom_sii_risk...() plot
}
\examples{
sii_z_ex8_map
}
\keyword{datasets}
