% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_z_example2_data}
\alias{sii_z_example2_data}
\title{sii_z_example2_data}
\format{A data frame with columns:
\describe{
           \item{time}{numerical: a representation of an x value: x aesthetic could also be a normal numbering, or a longitude}
           \item{ratio}{numerical: solvency II ratio, a representation of an y value: y aesthetic could also be a lattitude}
           \item{description}{character: component of the structure}
           \item{id}{ number: }
           \item{value}{numerical: }
           \item{comparewithid}{numerical: }
         }}
\source{
loosely based on public SFCR report of a medium sized dutch life insurer
made from excelfile in https://github.com:vanzanden/ggsolvencyii : \cr
\code{
 data <-  readxl::read_xlsx(path = "xls/preparingtables.xlsx",
                           sheet = "ex2_data") %>%
           tidyr::gather(key = description,
                         value = value,
                         -id, -time, -ratio)
 sii_z_example1_data <-  data.frame(
                                     time = as.numeric(data$time),
                                     ratio = as.numeric(data$ratio),
                                     description = data$description,   # it has to be a factor !!
                                     value = as.numeric(data$value),
                                     id = data$id,
                                     comparewithid = data$comparewithid
                                     )
 }
}
\usage{
sii_z_example2_data
}
\description{
example dataset, in tidyverse format, with several instances of a completely  filled sf structure,
}
\examples{
sii_z_example2_data
}
\keyword{datasets}
