% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_z_example3_data}
\alias{sii_z_example3_data}
\title{sii_z_example3_data}
\format{A data frame with columns:
\describe{
           \item{time}{numerical: a representation of an x value: \cr The x aesthetic could also be a normal numbering, or a longitude}
           \item{ratio}{numerical: solvency II ratio, a representation of an y value: \cr y aesthetic could also be a lattitude}
           \item{description}{character: component of the structure}
           \item{id}{ number: }
           \item{value}{numerical: }
           \item{comparewithid}{numerical: }
         }}
\source{
based on public SFCR 2017 reports of a medium sized dutch life insurer in Oosterhout
  made from excelfile in \href{https://github.com/vanzanden/ggsolvencyii/tree/master/xls}{github.com/vanzanden} : \cr
\code{
dummy
 }
}
\usage{
sii_z_example3_data
}
\description{
example dataset, in tidyverse format, with several instances of a filled sf structure,
}
\examples{
sii_z_example3_data
}
\keyword{datasets}
