% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_z_example3_structure}
\alias{sii_z_example3_structure}
\title{sii_z_example3_structure}
\format{A data frame with columns:
\describe{
          \item{description}{string: the name of the component, or a diversification}
          \item{level}{string with format 'level(numeric)' [<point> 'sublevel'] ['d'] } with d indicating diversification
          \item{childlevel}{string of the same format as level, indicating the level of which the components make up this item }
         }}
\source{
determined based on specification of EIOPA standard formula Solvency II structure

\cr import from accompanying excel:\cr
\code{
  sii_z_example3_structure <-   readxl::read_xlsx(path = "xls/preparingtables.xlsx",
                                               sheet = "ex3_struct",
                                               col_types = c(rep("text", 3),rep("skip", 2)))
 }
}
\usage{
sii_z_example3_structure
}
\description{
a standard formula structure (as of 2016), with descriptions in English
}
\examples{
sii_z_example3_structure
}
\keyword{datasets}
