% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_z_example4_outline}
\alias{sii_z_example4_outline}
\title{sii_z_example4_outline}
\format{An object of class \code{data.frame} with 18 rows and 7 columns.}
\source{
determined based on specification of EIOPA standard formula Solvency II structure
\code{
dummy
   data$levelordescription <- as.character(data$levelordescription);
   data$outline1 <- as.logical(data$outline1);
   data$outline2 <- as.logical(data$outline2);
   data$outline3 <- as.logical(data$outline3);
   data$outline4 <- as.logical(data$outline4);
   data$outline11 <- as.logical(data$outline11);
   data$outline13 <- as.logical(data$outline13);

   sii_outline_sf16_eng <- as.data.frame(data)
}
}
\usage{
sii_z_example4_outline
}
\description{
is a table for geom_solviioutline indicating which outlines of each circle segment should be shown, specified per level, as defined in the structure. A circle segment is defined by ri (radius inner), ro (radius outer), db (degrees beginning, based on a compas, with 0 degrees pointing up from the center) and de (degrees end). In this example file the components m_equity and l_lapse are individual described, overruling their standard level outline description
}
\details{
This example file highlights the posibilities to override the levelbased outlines for items (i.e. individual risks(accumulations)).
}
\examples{
sii_z_example4_outline
}
\keyword{datasets}
