% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_z_example4_outline_exceptions}
\alias{sii_z_example4_outline_exceptions}
\title{sii_z_example4_outline_exceptions}
\format{A data frame with columns:
\describe{
           \item{levelordescription}{string with format 'level(numeric)' <point> 'sublevel' 'd'  with d indicating diversification}
           \item{outline1}{logical: the radial line between ri and ro, for db }
           \item{outline2}{logical: the outer line between db and de, for ro }
           \item{outline3}{logical: the radial line between ro and ri, for de }
           \item{outline4}{logical: the inner line between de and db, for ri.}
           \item{outline11}{NOT YET IMPLEMENTED: logical: analogue to outline1, but but only for the FIRST plotted component of that level. The components in each  level are plotted clockwise}
           \item{outline13}{NOT YET IMPLEMENTED: logical: analogue to outline3, but but only for the LAST plotted component of that level. }
         }}
\source{
determined based on specification of EIOPA standard formula Solvency II structure
}
\usage{
sii_z_example4_outline_exceptions
}
\description{
is a table for geom_solviioutline indicating which outlines of each circle segment should be shown, specified per level, as defined in the structure. A circle segment is defined by ri (radius inner), ro (radius outer), db (degrees beginning, based on a compas, with 0 degrees pointing up from the center) and de (degrees end). In this example file the components m_equity and l_lapse are individual described, overruling their standard level outline description
}
\examples{
sii_z_example4_outline_exceptions
}
\keyword{datasets}
