% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101_geomandstat.R
\name{stat_solvii}
\alias{stat_solvii}
\title{stat_solvii}
\usage{
stat_solvii(mapping = NULL, data = NULL, geom = "solvii",
  position = "identity", show.legend = TRUE, inherit.aes = TRUE,
  na.rm = FALSE, levelmax = 99, structuredf = sii_structure_sf16_eng,
  outlinedf = sii_outline_sf16_eng, maxscrvalue = NULL,
  levelonedescription = "SCR", aggregatesuffix = "_other", scalingx = 1,
  scalingy = 1, fullstructure = FALSE, rotationdegrees = NULL,
  rotationdescription = NULL, ...)
}
\arguments{
\item{mapping}{required aes(thetics) : x (i.e. time, longitude, integer), y (i.e SCR ratio, lattitude), id, description (), value}

\item{data}{the dataset in tidyverse format (description as a factor)}

\item{geom}{the default is geom_solvii}

\item{position}{standard ggplot function}

\item{show.legend}{standard ggplot function}

\item{inherit.aes}{standard ggplot function}

\item{na.rm}{standard ggplot function}

\item{levelmax}{default = 99: a positive integer or a dataframe with columns 'level' and 'levelmax'. The maximum amount of items in a certain level to be plotted. The smallest items are combined}

\item{structuredf}{default = structure_sf_eng: a representation of the structure, columns description (chr), level (chr), childlevel (chr). description in dataset needs to match the desripton in this file. see ?structure_sf_eng}

\item{outlinedf}{default = sii_outlinedf_sf16_eng}

\item{maxscrvalue}{default = NULL: the scale of the different plot elements, is (by surfacearea) is by default measured to the largest element, this can be overridden by this parameter}

\item{levelonedescription}{default = "SCR": the description of level 1, this name is needed before integration of data and structure.}

\item{aggregatesuffix}{default = "_other" When a certain level contains more items than specified by levelmax (an integer or dataframe (level, levelmax)) the smallest items are combined. The description is of the motherlevel with a suffix}

\item{scalingx}{default = 1: for plots where units in x and y are different distortion can occur. This parameter scales only in x-direction}

\item{scalingy}{default = 1: for plots where units in x and y are different distortion can occur. This parameter scales only in y-direction}

\item{fullstructure}{default = FALSE: an indication if the legend has to show all possible components of the structure, or only those present in the dataset}

\item{rotationdegrees}{default = NULL: the fixed amount of degrees of which each item is rotated (as in a compas, -90 is a quarter rotation anti-clockwise), additive to possible rotation to description}

\item{rotationdescription}{default = NULL: which item determines the rotation of the plot, so that this item is}

\item{...}{ellipsis}
}
\value{
a ggplot object
}
\description{
stat_solvii
}
