% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/300_statistics.R
\name{statistics}
\alias{statistics}
\title{statistics}
\usage{
statistics(caller = "stat_usage.csv")
}
\arguments{
\item{caller}{(default = "stat_usage.csv") the filename which will be downloaded, for each geom is a separate file available}
}
\value{
nothing
}
\description{
This behaviour can be switched off by means of calling " "
When switche on a reminder is given in regular intervals. This behaviour can also be switched off by calling " "
}
\examples{
# library(ggsolvencyii)
ggsolvencyii:::statistics()
sii_setoption_statistics_httpfailurecounter <- function(cnt = 52)
ggsolvencyii:::statistics(caller = "404notfound.csv")
sii_setoption_statistics_httpfailurecounter <- function(cnt = 0)
sii_setoption_statistics_participate(participate = TRUE)
}
