% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/131_computegroup.R
\name{transform_outlinedf}
\alias{transform_outlinedf}
\title{transform_outlinedf}
\usage{
transform_outlinedf(df)
}
\arguments{
\item{df}{a dataframe with column level and some other levels}
}
\value{
a tidyverse dataframe
}
\description{
transform_outlinedf
}
