% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/111_setupparams.R
\name{fn_constructionplotdetails}
\alias{fn_constructionplotdetails}
\title{fn_constructionplotdetails}
\usage{
fn_constructionplotdetails(structure)
}
\arguments{
\item{structure}{the structure dataframe}
}
\value{
a dataframe columns levelordescription, surface, outline1 .. 11, with TRUE for all fields
}
\description{
When no plotdetails table is given then a default table is constructed based on the current structure.
}
