% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/121_setupdata_risksurface.R
\name{fn_structure_data_integration}
\alias{fn_structure_data_integration}
\title{fn_structure_data_integration combines data and expanded structure, calculation aggregated items and removing lines for which no aggregation is neccessary.}
\usage{
fn_structure_data_integration(expandedstructure, data)
}
\arguments{
\item{expandedstructure}{result of fn_structure_expansion()}

\item{data}{the data send by the geom_class, the result of the integration of user supplied data and expanded structure .}
}
\value{
data
}
\description{
fn_structure_data_integration combines data and expanded structure, calculation aggregated items and removing lines for which no aggregation is neccessary.
}
