% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/131_computegroup.R
\name{fn_transform_plotdetails}
\alias{fn_transform_plotdetails}
\title{fn_transform_plotdetails}
\usage{
fn_transform_plotdetails(df, siiparams)
}
\arguments{
\item{df}{a dataframe with column level and some other levels}

\item{outline}{boolean, TRUE if calling geom is geom_sii_riskoutline , only one boolean can be set to TRUE}

\item{surface}{boolean, TRUE if calling geom is geom_sii_risksurface , only one boolean can be set to TRUE}
}
\value{
a tidyverse dataframe
}
\description{
fn_transform_plotdetails
}
