% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/136_computegroup_subA.R
\name{fnRotation}
\alias{fnRotation}
\title{fnRotation}
\usage{
fnRotation(df, siiparams)
}
\arguments{
\item{df}{dummy text}

\item{siiparams}{dummy text}
}
\value{
a dataframe with adjusted db and de columns
}
\description{
fnRotation
}
