% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/123_setupdataconnection.R
\name{fnSetupdata_connection}
\alias{fnSetupdata_connection}
\title{fnSetupdata_connection}
\usage{
fnSetupdata_connection(data)
}
\arguments{
\item{data}{the data send by the geom_class, due to the simple nature of the desired plot no structure information is needed.}
}
\value{
an adjusted version of input dataframe data
}
\description{
fnSetupdata_connection
}
