% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101_geomandstat.R
\name{geom_solvii}
\alias{geom_solvii}
\title{geom_solvii returns a ggplot2 object with filled, concentric circle(part)s}
\usage{
geom_solvii(data = NULL, mapping = NULL, stat = "solvii",
  structuredf = sii_structure_sf16_eng, levelmax = 99,
  aggregatesuffix = "_other", maxscrvalue = NULL, scalingx = 1,
  scalingy = 1, rotationdegrees = NULL, rotationdescription = NULL,
  squared = FALSE, position = "identity", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{data}{the dataset in tidyverse format (column 'description' as a factor)}

\item{mapping}{required aes(thetics) : x (i.e. time, longitude), y (i.e SCR ratio, lattitude), id, description (), value}

\item{stat}{default stat is statSolvii}

\item{structuredf}{(dataframe: default = sii_structure_sf16_eng): a representation of the structure, columns are description (chr), level (chr), childlevel (chr). In the standard formula structure, SCR has level 1, with childlevel 2. This means it consists of all datalines with level == 2, ie. BSCR, operational and Adjustment-LACDT. lines in the dataseet with a suffix "d" behind the levelnumber are diversifacation items. As of now these are not used in any calculation. the values in column description in the dataset need to match the desription in this file.  see ?sii_structure_sf16_eng. The package contains also a file sii_structure_sf16_nld with Dutch terms in descripton column.}

\item{levelmax}{(integer or dataframe, default = 99): a positive integer or a dataframe with columns 'level' and 'levelmax'. The maximum amount of items in a certain level to be plotted. The smallest items are combined to one item. In the case level consisting of 7 items has a levelmax of 5 this results in 4 seperate items and one grouped item.}

\item{aggregatesuffix}{(string, default = "_other") When a certain level contains more items than specified by levelmax  the smallest items are combined. The description is of the name of the one higher level (lower number) with a suffix}

\item{maxscrvalue}{(optional, double, default = NULL): the scale of the different plot elements is is by default measured to the largest level 1 element (i.e. SCR) in the dataset, this can be overridden by this parameter for example when combining several plots}

\item{scalingx}{(optional, positive value ,default = 1): for plots where units in x and y are different in magnitude distortion can occur. This parameter scales only in x-direction}

\item{scalingy}{(optional, positive value ,default = 1): for plots where units in x and y are different in magnitude distortion can occur. This parameter scales only in y-direction}

\item{rotationdegrees}{(optional, integer, -360 to 360, default = NULL): when given, the fixed amount of degrees (positive is clockwise) of which each item is rotated (as in a compas, -90 is a quarter rotation anti-clockwise), additive to possible rotation to description}

\item{rotationdescription}{(optional, string, default = NULL): default the orientation of the lower level (higher number) circles is based on the structure. When this parameter is not NULL then the circles are rotated in such a way that the indicated item lies in the "north-east" part of the circle.}

\item{squared}{(optional, boolean, default = FALSE): when set to TRUE plot returns a square representation. Compared with a circle representation of the same data the heigth and width of the square are smaller than the radius of the circle. Segments which fall in the corner parts of the square are smaller than equalsized part which fall in the vertical or horizontal parts of the square.
#'}

\item{position}{standard ggplot function}

\item{na.rm}{standard ggplot function}

\item{show.legend}{standard ggplot function}

\item{inherit.aes}{standard ggplot function}

\item{...}{ellipsis, a standard R parameter}
}
\value{
a ggplot object
}
\description{
geom_solvii returns a ggplot2 object with filled, concentric circle(part)s
}
