% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/101_geomandstat.R
\name{geom_solviiconnection}
\alias{geom_solviiconnection}
\title{geom_solviiconnection}
\usage{
geom_solviiconnection(data = NULL, mapping = NULL,
  stat = "solviiconnection", position = "identity", na.rm = FALSE,
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{data}{the dataset in tidyverse format (description as a factor)}

\item{mapping}{required aes(thetics) : x (i.e. time, longitude, integer), y (i.e SCR ratio, lattitude), id, description (), value}

\item{stat}{default stat is statSolvii}

\item{position}{standard ggplot function}

\item{na.rm}{standard ggplot function}

\item{show.legend}{standard ggplot function}

\item{inherit.aes}{standard ggplot function}

\item{...}{ellipsis}
}
\value{
a ggplot object
}
\description{
geom_solviiconnection
}
