% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/200_ggsolvencyii_debug.R
\name{sii_debug}
\alias{sii_debug}
\title{sii_debug}
\usage{
sii_debug(data_descriptionvector,
  structure = ggsolvencyii::sii_structure_sf16_eng,
  aggregatesuffix = "other", levelmax = NULL, plotdetails = NULL,
  fillcolors = NULL, edgecolors = NULL)
}
\arguments{
\item{structure}{(no default) the structure dataframe}

\item{aggregatesuffix}{(optional, default = "_other"):}

\item{levelmax}{(optional, no default): the levelmax dataframe, when not filled the expanded structure assumes levelmax was set to 99 for all levels.}

\item{plotdetails}{(optional, no default): the plotdetails dataframe}

\item{fillcolors}{(optional, no default): fillcolor parameter (list with items "description" = "color" where color can be a name, #hexcode or other )}

\item{edgecolors}{(optional, no default): edgecolor parameter (list with items "description" = "color" where color can be a name, #hexcode or other )}

\item{data_descr}{(no default) the vector 'description' from the data}
}
\value{
prints two comparison tables and puts them in $debug_description and $debug_level
}
\description{
assists in finding level or description mismatches in a set parameter(tables)
}
