% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/200_ggsolvencyii_debug.R
\name{sii_debug}
\alias{sii_debug}
\title{sii_debug}
\usage{
sii_debug(data_descr, structure = ggesolvencii::sii_structure_sf16_eng,
  levelmax = ggesolvencii::sii_levelmax_995,
  outline = ggesolvencii::sii_outline_sf16_eng,
  fillcolor = ggsolvencyii::sii_x_fillcolors_sf16_eng,
  edgecolor = ggsolvencyii::sii_x_edgecolors_sf16_eng,
  aggregatesuffix = "other")
}
\arguments{
\item{data_descr}{(no default) the vector 'description' from the data}

\item{structure}{(default = sii_structure_sf16_eng) : the structure dataframe}

\item{levelmax}{(default = sii_levelmax_995): the levelmax dataframe}

\item{outline}{(default = sii_outline_sf16_eng): the outline dataframe}

\item{fillcolor}{(default = sii_x_fillcolors_sf16_eng): fillcolor parameter}

\item{edgecolor}{(default = sii_x_fillcolors_sf16_eng): edgecolor parameter}

\item{aggregatesuffix}{(default = "_other"):}
}
\value{
prints two comparison tables and puts them in $debug_description and $debug_level
}
\description{
assists in finding level or description mismatches in a set parameter(tables)
}
