% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_levelmax_sf16_995}
\alias{sii_levelmax_sf16_995}
\title{sii_levelmax_sf16_995}
\format{A data frame with columns:
\describe{
          \item{level}{string with format \emph{'level(numeric)' [<point> 'sublevel(numeric)'] ['d']} i.e. 1, 2.01, 3.1d where d indicates a diversification component}
          \item{levelmax}{positive integer indicating the total number of components to be shown in a level, consisting of the x-1 largest components and the remaining components combined in one other}
         }}
\source{
made from excel-file in \href{https://github.com/vanzanden/ggsolvencyii/tree/master/_data_helperfiles}{github.com/vanzanden}, from there transfered to R environment with code in \code{preparingtables.R}.
}
\usage{
sii_levelmax_sf16_995
}
\description{
maximum number of items in a level, 99 items for levels 1-3
(SCR, BSCR/OR/ADJ, Market/life/... risks) and 5 items for lower level risk
(equity/longevity/...) in a standard formula structure.
}
\examples{
installedtable <- sii_levelmax_sf16_995
installedtable
}
\keyword{datasets}
