% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_outline_sf16_eng}
\alias{sii_outline_sf16_eng}
\title{sii_outline_sf16_eng}
\format{A data frame with columns:
\describe{
           \item{levelordescription}{string with format 'level(numeric)' <point> 'sublevel' 'd'  with d indicating diversification}
           \item{outline1}{boolean (logical): the radial line between ri and ro, for db }
           \item{outline2}{boolean (logical): the outer line between db and de, for ro }
           \item{outline3}{boolean (logical): the radial line between ro and ri, for de }
           \item{outline4}{boolean (logical): the inner line between de and db, for ri.}
           \item{outline11}{NOT YET IMPLEMENTED: boolean (logical): analogue to outline1, but but only for the first plotted component of that level. The components in each  level are plotted clockwise}
           \item{outline13}{NOT YET IMPLEMENTED: boolean (logical): analogue to outline3, but but only for the last plotted component of that level. }
         }}
\source{
determined based on specification of EIOPA standard formula Solvency II structure
\cr import from accompanying excel:\cr
\code{
 data <-  readxl::read_xlsx(path = "xls/preparingtables.xlsx",
                           sheet = "outline_sf16_eng")
   data$levelordescription <- as.character(data$levelordescription)
   data$outline1 <- as.logical(data$outline1)
   data$outline2 <- as.logical(data$outline2)
   data$outline3 <- as.logical(data$outline3)
   data$outline4 <- as.logical(data$outline4)
   data$outline11 <- as.logical(data$outline11)
   data$outline13 <- as.logical(data$outline13)

   sii_outline_sf16_eng <- as.data.frame(data)
}
}
\usage{
sii_outline_sf16_eng
}
\description{
A table for \code{\link{geom_solviioutline}} indicating which outlines of each item should be shown, specified per level and/or description. the latter overrule the former. \cr
when defining an item (or the \code{squared = TRUE} transformation) 4 lines can be distinguised, a radialline going outwards, a circle segment (clockwise), a radialline going inwards, a circle segment (counterclockwise). These are numbered as outline1 to outline4.
}
\examples{
sii_outline_sf16_eng
}
\keyword{datasets}
