% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_z_ex4_levelmax}
\alias{sii_z_ex4_levelmax}
\title{sii_z_ex4_levelmax}
\format{A data frame with columns:}
\source{
made from excel-file in \href{https://github.com/vanzanden/ggsolvencyii/tree/master/_data_helperfiles}{github.com/vanzanden}, from there transfered to R environment with code in \code{preparingtables.R}.
}
\usage{
sii_z_ex4_levelmax
}
\description{
maximum number of items in a level, 99 items for levels 1-3 (SCR, BSCR/OR/ADJ, Market/life/... risks) and 3 items for lowerlevel risk (equity/longevity/...)
}
\examples{
installedtable <- sii_z_ex4_levelmax
installedtable
}
\keyword{datasets}
