% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_z_example3_levelmax}
\alias{sii_z_example3_levelmax}
\title{sii_z_example3_levelmax}
\format{A data frame with columns:}
\source{
\code{
dummy
}
}
\usage{
sii_z_example3_levelmax
}
\description{
tekst
}
\examples{
sii_z_example3_levelmax
}
\keyword{datasets}
