% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sii_z_example3_levelmax}
\alias{sii_z_example3_levelmax}
\title{sii_z_example3_levelmax}
\format{A data frame with columns:}
\source{
made from excelfile in \href{https://github.com/vanzanden/ggsolvencyii/tree/master/xls}{github.com/vanzanden} : \cr
\code{
dummy
 data <- readxl::read_xlsx(path = "xls/preparingtables.xlsx",sheet = "ex3_levelmax");
 sii_z_example3_levelmax <- data.frame(level = as.character(data$level), levelmax = as.numeric(data$levelmax))
}
}
\usage{
sii_z_example3_levelmax
}
\description{
tekst
}
\examples{
sii_z_example3_levelmax
}
\keyword{datasets}
